Projects menu for creating and searching projects separated to own directory
