--- client/src/components/ChatBox/CurrentChatWindow/Profile/index.js
@@ -1,17 +1,0 @@
-import { useState } from 'react'
-import SettingsPage from '../../../../../pages/SettingsPage'
-import TransparentBackground from '../../../../CommonComponents/TransparentBackground'
-
-const Profile = () => {
-    const [profile, setProfile] = useState(true)
-
-    if (!profile) return null
-    return (
-        <>
-            <TransparentBackground closeOpenedWindows={() => setProfile(false)} />
-            <SettingsPage />
-        </>
-    )
-}
-
-export default Profile
