+++ client/src/components/ChatBox/CurrentChatWindow/UserMenu/index.js
@@ -1,8 +1,8 @@
 import styles from './index.module.css'
 import UserNav from '../UserNav'
-import CreateProject from '../CreateProject'
-import SearchProjects from '../SearchProjects'
-import PendingProjects from '../PendingProjects'
+import CreateProject from '../../../ProjectsMenu/CreateProject'
+import SearchProjects from '../../../ProjectsMenu/SearchProjects'
+import PendingProjects from '../../../ProjectsMenu//PendingProjects'
 import { useContext } from 'react'
 import { MessagesContext } from '../../../../context/MessagesContext'
 import PersonalSettingsColumn from '../../../ProfileSettings/PersonalSettingsColumn'
